.Core: Fix #13 (MonoLive killing XmlHelper); .Editor: Add speech line autocompletion.
The fix for MonoLive killing XmlHelper is somewhat aggressive, but at least XmlHelper is working in
the MonoKickstart (Steam) version of FEZ again. Unfortunately it requires an update to the Travis
CI build list, mod.bat / mod.sh (new assembly to be patched) and MonoMod (replace / remove feature).

Speech line autocompletion is currently suffering from a horrible clipping / scissoring issue where
the scissor of the list container widget clips the background of the text field children (the list)
while keeping the font working.